refactor(cel): extract cel-contracts crate (PR1a.0 — mechanical, no behaviour change)
cel-cortex historically depended on cel-planner because shared boundary types
(PlannedAction, Step, NextMove, PlanProducer trait, …) lived in cel-planner.
That made the platform layer depend on a planner implementation, which is the
wrong direction once we want planners to be peers under CEL.

This PR extracts those types into a new cel-contracts crate so cortex,
runner, and any future planner runtime (LangGraph, Mastra, Codex, in-house)
can speak the contract without depending on each other.

What moved to cel-contracts:
  - PlannedAction + CellWrite (action contract)
  - Step, StepKind, StepResult (execution contract)
  - RuntimeCaps, RunLimits, AttemptRecord (runtime/run-state contract)
  - NextMove, GoalOutcome, FailureReport (planner output / run outcome)
  - PlanProducer trait + DoneVerdict (the contract itself)

What stays in cel-planner:
  - Plan, SubGoal (legacy multi-step plan types)
  - Planner, LlmPlanProducer, CanonicalPlanProducer (impls)
  - prompt::*, distiller::*, history::*, signals::*, decompose::*
  - PlannedStep, PlannedPlan, NotebookWrite, ContextDetail/Tier, GoalConfig,
    ProgressAssessment, PlannerEvent, StepRecord (planner-internal)
  - PlannerError, PlannerBackend

cel-contracts depends only on cel-context + serde + async-trait — no
cel-llm, no prompt logic, no planner internals. That keeps the boundary
crate honest and small.

Backward compat: cel-planner re-exports the moved types via
`pub use cel_contracts::{...}`. Existing callers continue to work via the
old `cel_planner::PlannedAction` etc. paths. New code should import from
`cel_contracts` directly.

Dep graph after this PR:

    cel-context     (foundational types)
       ↑
    cel-contracts   (boundary types — depends on cel-context)
       ↑
    cel-cortex      (depends on context + contracts; NO LONGER on cel-planner)
    cel-planner     (depends on context + contracts; peer with cortex)
       ↑
    cel-goal-runner (wires cortex + planner together)

cel-cortex's cel-planner dep is removed entirely. cel-goal-runner and
cel-napi keep both deps (they still use cel-planner for the LlmPlanProducer
impl) but their boundary-type imports now come from cel-contracts.

Also adds COGNITION_LAYER_PLAN.md — the PR1a..PR4 design doc that
documents this extraction as PR1a.0 and the upcoming PlanningView work.

Tests: 195 tests pass across cel-contracts, cel-planner, cel-cortex,
cel-goal-runner. cargo check --workspace --all-targets clean.
