+++ cel/cel-cortex/Cargo.toml
@@ -11,7 +11,7 @@ cel-context = { workspace = true }
 cel-audio = { workspace = true }
 cel-accessibility = { workspace = true }
 cel-input = { workspace = true }
-cel-planner = { workspace = true }
+cel-contracts = { workspace = true }
 cel-signals = { workspace = true }
 cel-network = { workspace = true }
 cel-cdp = { workspace = true }
