+++ cel/cel-goal-runner/Cargo.toml
@@ -15,6 +15,7 @@ cel-cortex = { workspace = true }
 # look at. Must stay on wherever the runner runs live.
 cel-display = { workspace = true, features = ["xcap"] }
 cel-input = { workspace = true }
+cel-contracts = { workspace = true }
 cel-planner = { workspace = true }
 cel-llm = { workspace = true }
 cel-store = { workspace = true }
