+++ cel/cel-napi/Cargo.toml
@@ -21,6 +21,7 @@ cel-network = { workspace = true }
 cel-audio = { workspace = true }
 cel-vision = { workspace = true }
 cel-store = { workspace = true }
+cel-contracts = { workspace = true }
 cel-planner = { workspace = true }
 cel-cdp = { workspace = true }
 cel-cortex = { workspace = true }
