+++ cel/cel-napi/src/goal_runner.rs
@@ -18,9 +18,10 @@ use cel_goal_runner::{
     resolve_runtime_backend, CanonicalGoalRunner, CortexStepExecutor, GoalConfig, RuntimeBackend,
     StepExecutor,
 };
-use cel_planner::{
-    AttemptRecord, GoalOutcome, LlmPlanProducer, PlanProducer, RunLimits, RuntimeCaps, Step,
+use cel_contracts::{
+    AttemptRecord, GoalOutcome, PlanProducer, RunLimits, RuntimeCaps, Step,
 };
+use cel_planner::LlmPlanProducer;
 use cellar_worker::{SubmitGoalRequest, WorkerClient};
 use serde::Deserialize;
